requirements.txt
discord.py
psutil
python-dotenv
bitcoin
